+++ NEWS.md
@@ -1,5 +1,8 @@
 # dplyr 0.2.0.9000
 
+* `regroup()` is deprecated. Please use the more flexible `group_by_()`
+  instead.
+
 * You can now program with dplyr - every function that does non-standard
   evaluation (NSE) has a standard evaluation (SE) version ending in `_`.
   This is powered by the new lazyeval package which provides all the tools